Struct aws_sdk_mq::client::fluent_builders::CreateBroker

pub struct CreateBroker<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateBroker<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateBrokerOutput, SdkError<CreateBrokerError>> where
R::Policy: SmithyRetryPolicy<CreateBrokerInputOperationOutputAlias, CreateBrokerOutput, CreateBrokerError, CreateBrokerInputOperationRetryAlias>,

pub fn authentication_strategy(self, inp: AuthenticationStrategy) -> Self

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

pub fn set_authentication_strategy(
self,
input: Option<AuthenticationStrategy>
) -> Self

pub fn auto_minor_version_upgrade(self, inp: bool) -> Self

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot. Set to true by default, if no value is specified.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn broker_name(self, inp: impl Into<String>) -> Self

Required. The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

pub fn set_broker_name(self, input: Option<String>) -> Self

pub fn configuration(self, inp: ConfigurationId) -> Self

A list of information about the configuration.

pub fn set_configuration(self, input: Option<ConfigurationId>) -> Self

pub fn creator_request_id(self, inp: impl Into<String>) -> Self

The unique ID that the requester receives for the created broker. Amazon MQ passes your ID with the API action. Note: We recommend using a Universally Unique Identifier (UUID) for the creatorRequestId. You may omit the creatorRequestId if your application doesn't require idempotency.

pub fn set_creator_request_id(self, input: Option<String>) -> Self

pub fn deployment_mode(self, inp: DeploymentMode) -> Self

Required. The broker's deployment mode.

pub fn set_deployment_mode(self, input: Option<DeploymentMode>) -> Self

pub fn encryption_options(self, inp: EncryptionOptions) -> Self

Encryption options for the broker. Does not apply to RabbitMQ brokers.

pub fn set_encryption_options(self, input: Option<EncryptionOptions>) -> Self

pub fn engine_type(self, inp: EngineType) -> Self

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

pub fn set_engine_type(self, input: Option<EngineType>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

Required. The broker engine's version. For a list of supported engine versions, see Supported engines.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn host_instance_type(self, inp: impl Into<String>) -> Self

Required. The broker's instance type.

pub fn set_host_instance_type(self, input: Option<String>) -> Self

pub fn ldap_server_metadata(self, inp: LdapServerMetadataInput) -> Self

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

pub fn set_ldap_server_metadata(
self,
input: Option<LdapServerMetadataInput>
) -> Self

pub fn logs(self, inp: Logs) -> Self

Enables Amazon CloudWatch logging for brokers.

pub fn set_logs(self, input: Option<Logs>) -> Self

pub fn maintenance_window_start_time(self, inp: WeeklyStartTime) -> Self

The parameters that determine the WeeklyStartTime.

pub fn set_maintenance_window_start_time(
self,
input: Option<WeeklyStartTime>
) -> Self

pub fn publicly_accessible(self, inp: bool) -> Self

Enables connections from applications outside of the VPC that hosts the broker's subnets. Set to false by default, if no value is provided.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn security_groups(self, inp: impl Into<String>) -> Self

Appends an item to SecurityGroups.

To override the contents of this collection use set_security_groups.

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn storage_type(self, inp: BrokerStorageType) -> Self

The broker's storage type.

pub fn set_storage_type(self, input: Option<BrokerStorageType>) -> Self

pub fn subnet_ids(self, inp: impl Into<String>) -> Self

Appends an item to SubnetIds.

To override the contents of this collection use set_subnet_ids.

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones. If you specify more than one subnet, the subnets must be in different Availability Zones. Amazon MQ will not be able to create VPC endpoints for your broker with multiple subnets in the same Availability Zone. A SINGLE_INSTANCE deployment requires one subnet (for example, the default subnet). An ACTIVE_STANDBY_MULTI_AZ Amazon MQ for ActiveMQ deployment requires two subnets. A CLUSTER_MULTI_AZ Amazon MQ for RabbitMQ deployment has no subnet requirements when deployed with public accessibility. Deployment without public accessibility requires at least one subnet.

If you specify subnets in a shared VPC for a RabbitMQ broker, the associated VPC to which the specified subnets belong must be owned by your AWS account. Amazon MQ will not be able to create VPC endpoints in VPCs that are not owned by your AWS account.

pub fn set_subnet_ids(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

Create tags when creating the broker.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn users(self, inp: impl Into<User>) -> Self

Appends an item to Users.

To override the contents of this collection use set_users.

Required. The list of broker users (persons or applications) who can access queues and topics. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.
Amazon MQ for RabbitMQ
When you create an Amazon MQ for RabbitMQ broker, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

pub fn set_users(self, input: Option<Vec<User>>) -> Self