Module aws_sdk_mq::model

Modules

availability_zone

See AvailabilityZone

broker_engine_type

See BrokerEngineType

broker_instance

See BrokerInstance

broker_instance_option

See BrokerInstanceOption

broker_summary

See BrokerSummary

configuration

See Configuration

configuration_id

See ConfigurationId

configuration_revision

See ConfigurationRevision

configurations

See Configurations

encryption_options

See EncryptionOptions

engine_version

See EngineVersion

ldap_server_metadata_input

See LdapServerMetadataInput

ldap_server_metadata_output

See LdapServerMetadataOutput

logs

See Logs

logs_summary

See LogsSummary

pending_logs

See PendingLogs

sanitization_warning

See SanitizationWarning

user

See User

user_pending_changes

See UserPendingChanges

user_summary

See UserSummary

weekly_start_time

See WeeklyStartTime

Structs

AvailabilityZone

Name of the availability zone.

BrokerEngineType

Types of broker engines.

BrokerInstance

Returns information about all brokers.

BrokerInstanceOption

Option for host instance type.

BrokerSummary

Returns information about all brokers.

Configuration

Returns information about all configurations.

ConfigurationId

A list of information about the configuration.

Does not apply to RabbitMQ brokers.

ConfigurationRevision

Returns information about the specified configuration revision.

Configurations

Broker configuration information

EncryptionOptions

Does not apply to RabbitMQ brokers.

Encryption options for the broker.

EngineVersion

Id of the engine version.

LdapServerMetadataInput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker.

Does not apply to RabbitMQ brokers.

LdapServerMetadataOutput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker.

Logs

The list of information about logs to be enabled for the specified broker.

LogsSummary

The list of information about logs currently enabled and pending to be deployed for the specified broker.

PendingLogs

The list of information about logs to be enabled for the specified broker.

SanitizationWarning

Returns information about the XML element or attribute that was sanitized in the configuration.

User

A user associated with the broker. For RabbitMQ brokers, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

UserPendingChanges

Returns information about the status of the changes pending for the ActiveMQ user.

UserSummary

Returns a list of all broker users. Does not apply to RabbitMQ brokers.

WeeklyStartTime

The scheduled time period relative to UTC during which Amazon MQ begins to apply pending updates or patches to the broker.

Enums

AuthenticationStrategy

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

BrokerState

The broker's status.

BrokerStorageType

The broker's storage type.

EFS is not supported for RabbitMQ engine type.

ChangeType

The type of change pending for the ActiveMQ user.

DayOfWeek

DeploymentMode

The broker's deployment mode.

EngineType

The type of broker engine. Amazon MQ supports ActiveMQ and RabbitMQ.

SanitizationWarningReason

The reason for which the XML elements or attributes were sanitized.