Struct aws_sdk_mq::model::LdapServerMetadataOutput

#[non_exhaustive]
pub struct LdapServerMetadataOutput {
    pub hosts: Option<Vec<String>>,
    pub role_base: Option<String>,
    pub role_name: Option<String>,
    pub role_search_matching: Option<String>,
    pub role_search_subtree: bool,
    pub service_account_username: Option<String>,
    pub user_base: Option<String>,
    pub user_role_name: Option<String>,
    pub user_search_matching: Option<String>,
    pub user_search_subtree: bool,
}

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

hosts: Option<Vec<String>>

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

role_base: Option<String>

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

role_name: Option<String>

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

role_search_matching: Option<String>

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

role_search_subtree: bool

The directory search scope for the role. If set to true, scope is to search the entire subtree.

service_account_username: Option<String>

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

user_base: Option<String>

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

user_role_name: Option<String>

Specifies the name of the LDAP attribute for the user group membership.

user_search_matching: Option<String>

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

user_search_subtree: bool

The directory search scope for the user. If set to true, scope is to search the entire subtree.

Implementations

impl LdapServerMetadataOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture LdapServerMetadataOutput