Module aws_sdk_mq::operation

Structs

CreateBroker

Creates a broker. Note: This API is asynchronous.

To create a broker, you must either use the AmazonMQFullAccess IAM policy or include the following EC2 permissions in your IAM policy.

ec2:CreateNetworkInterface

This permission is required to allow Amazon MQ to create an elastic network interface (ENI) on behalf of your account.

ec2:CreateNetworkInterfacePermission

This permission is required to attach the ENI to the broker instance.

ec2:DeleteNetworkInterface

ec2:DeleteNetworkInterfacePermission

ec2:DetachNetworkInterface

ec2:DescribeInternetGateways

ec2:DescribeNetworkInterfaces

ec2:DescribeNetworkInterfacePermissions

ec2:DescribeRouteTables

ec2:DescribeSecurityGroups

ec2:DescribeSubnets

ec2:DescribeVpcs

For more information, see Create an IAM User and Get Your AWS Credentials and Never Modify or Delete the Amazon MQ Elastic Network Interface in the Amazon MQ Developer Guide.

CreateConfiguration

Creates a new configuration for the specified configuration name. Amazon MQ uses the default configuration (the engine type and version).

CreateTags

Add a tag to a resource.

CreateUser

Creates an ActiveMQ user.

DeleteBroker

Deletes a broker. Note: This API is asynchronous.

DeleteTags

Removes a tag from a resource.

DeleteUser

Deletes an ActiveMQ user.

DescribeBroker

Returns information about the specified broker.

DescribeBrokerEngineTypes

Describe available engine types and versions.

DescribeBrokerInstanceOptions

Describe available broker instance options.

DescribeConfiguration

Returns information about the specified configuration.

DescribeConfigurationRevision

Returns the specified configuration revision for the specified configuration.

DescribeUser

Returns information about an ActiveMQ user.

ListBrokers

Returns a list of all brokers.

ListConfigurationRevisions

Returns a list of all revisions for the specified configuration.

ListConfigurations

Returns a list of all configurations.

ListTags

Lists tags for a resource.

ListUsers

Returns a list of all ActiveMQ users.

RebootBroker

Reboots a broker. Note: This API is asynchronous.

UpdateBroker

Adds a pending configuration change to a broker.

UpdateConfiguration

Updates the specified configuration.

UpdateUser

Updates the information for an ActiveMQ user.