Struct aws_sdk_mq::output::DescribeBrokerOutput

#[non_exhaustive]
pub struct DescribeBrokerOutput {

    pub authentication_strategy: Option<AuthenticationStrategy>,
    pub auto_minor_version_upgrade: bool,
    pub broker_arn: Option<String>,
    pub broker_id: Option<String>,
    pub broker_instances: Option<Vec<BrokerInstance>>,
    pub broker_name: Option<String>,
    pub broker_state: Option<BrokerState>,
    pub configurations: Option<Configurations>,
    pub created: Option<Instant>,
    pub deployment_mode: Option<DeploymentMode>,
    pub encryption_options: Option<EncryptionOptions>,
    pub engine_type: Option<EngineType>,
    pub engine_version: Option<String>,
    pub host_instance_type: Option<String>,
    pub ldap_server_metadata: Option<LdapServerMetadataOutput>,
    pub logs: Option<LogsSummary>,
    pub maintenance_window_start_time: Option<WeeklyStartTime>,
    pub pending_authentication_strategy: Option<AuthenticationStrategy>,
    pub pending_engine_version: Option<String>,
    pub pending_host_instance_type: Option<String>,
    pub pending_ldap_server_metadata: Option<LdapServerMetadataOutput>,
    pub pending_security_groups: Option<Vec<String>>,
    pub publicly_accessible: bool,
    pub security_groups: Option<Vec<String>>,
    pub storage_type: Option<BrokerStorageType>,
    pub subnet_ids: Option<Vec<String>>,
    pub tags: Option<HashMap<String, String>>,
    pub users: Option<Vec<UserSummary>>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

authentication_strategy: Option<AuthenticationStrategy>

The authentication strategy used to secure the broker. The default is SIMPLE.

auto_minor_version_upgrade: bool

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

broker_arn: Option<String>

The broker's Amazon Resource Name (ARN).

broker_id: Option<String>

The unique ID that Amazon MQ generates for the broker.

broker_instances: Option<Vec<BrokerInstance>>

A list of information about allocated brokers.

broker_name: Option<String>

The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

broker_state: Option<BrokerState>

The broker's status.

configurations: Option<Configurations>

The list of all revisions for the specified configuration.

created: Option<Instant>

The time when the broker was created.

deployment_mode: Option<DeploymentMode>

The broker's deployment mode.

encryption_options: Option<EncryptionOptions>

Encryption options for the broker. Does not apply to RabbitMQ brokers.

engine_type: Option<EngineType>

The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

engine_version: Option<String>

The broker engine's version. For a list of supported engine versions, see Supported engines.

host_instance_type: Option<String>

The broker's instance type.

ldap_server_metadata: Option<LdapServerMetadataOutput>

The metadata of the LDAP server used to authenticate and authorize connections to the broker.

logs: Option<LogsSummary>

The list of information about logs currently enabled and pending to be deployed for the specified broker.

maintenance_window_start_time: Option<WeeklyStartTime>

The parameters that determine the WeeklyStartTime.

pending_authentication_strategy: Option<AuthenticationStrategy>

The authentication strategy that will be applied when the broker is rebooted. The default is SIMPLE.

pending_engine_version: Option<String>

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

pending_host_instance_type: Option<String>

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

pending_ldap_server_metadata: Option<LdapServerMetadataOutput>

The metadata of the LDAP server that will be used to authenticate and authorize connections to the broker after it is rebooted.

pending_security_groups: Option<Vec<String>>

The list of pending security groups to authorize connections to brokers.

publicly_accessible: bool

Enables connections from applications outside of the VPC that hosts the broker's subnets.

security_groups: Option<Vec<String>>

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

storage_type: Option<BrokerStorageType>

The broker's storage type.

subnet_ids: Option<Vec<String>>

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones.

tags: Option<HashMap<String, String>>

The list of all tags associated with this broker.

users: Option<Vec<UserSummary>>

The list of all broker usernames for the specified broker.

Implementations

impl DescribeBrokerOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeBrokerOutput